dagster/requirements.txt
# Dagster Dependencies
dagster>=1.6.0
dagster-webserver>=1.6.0
dagster-dbt>=0.22.0
dagster-postgres>=0.22.0
dagster-gcp>=0.22.0

# dbt 
dbt-core>=1.7.0
dbt-postgres>=1.7.0
dbt-bigquery>=1.7.0

# Google Cloud
google-cloud-bigquery>=3.0.0
google-cloud-storage>=2.0.0
google-cloud-dataproc>=5.0.0

# Utilities
pandas>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
